refactor: 가중치 수정 및 평점 Float 타입으로 수정

+++ src/main/java/store/mybooks/resource/book/dto/response/BookBriefResponse.java
@@ -27,7 +27,7 @@ public class BookBriefResponse {
 
     private String name;
 
-    private Double rate;
+    private Float rate;
 
     private Integer cost;
